Move logic tests into __tests__
